Superglobal 'REMOTE_USER' susceptible to XSS
The $_SERVER['REMOTE_USER'] isn't sanitized prior to being echoed out which results in a XSS (though this would be difficult for an attacker to reproduce due to needing access to the users file).

To reproduce, add new user to: /usr/local/nagiosxi/etc/htpasswd.users
with username: <script>alert(document.cookie)</script> and reload the page resulting in the cookie being displayed.